+++ lib/mock.dart
@@ -78,3 +78,5 @@ List searchHistory = [];
 
 const double userLat = 46.348838;
 const double userLng = 48.029816;
+
+List<Sight> sortedByRadius = [];
